requirements.txt
fastapi
uvicorn
genai
python-dotenv
mysql-connector-python
tiktoken
uuid
google-generativeai